refactor: split cmd_store.go along its three concerns
cmd_store.go was 770 lines and the largest file in the package by 250. It held
three things: the five store subcommands, the interactive workflow for choosing
a store's encryption method, and the reachability/initialization checks.

The last two are not `store` command code. `profile new` and `setup` both call
checkOrInitStore and promptEncryptionConfig, so those workflows were living
inside a command file that happens to be where they were first written.

  cmd_store.go        366  the five subcommands and their flags
  store_encryption.go 230  choosing an encryption method, where secrets live
  store_health.go     213  reachable? initialized? plus the AWS SSO re-auth path

Nothing moved packages, so this changes no behaviour: all 33 declarations are
present exactly once, and every line of logic is byte-identical to what it
replaced — the only content difference is the two files' doc headers and their
import blocks.
