+++ AGENTS.md
@@ -43,7 +43,7 @@ Docker-based hermetic tests (MinIO store, SFTP source/store) are automatically s
 ### Package Layout
 
 - Root package (`client.go`, `repo.go`, `aliases.go`, `backup.go`, `query.go`, `retention.go`, `restore.go`, `check.go`) — the public `Client` API and the library entry point. Split by domain rather than held in one file: `client.go` is the `Client`, its options and the store chain; `repo.go` covers init, format upgrade and key slots; `aliases.go` holds the type re-exports; the rest map to operations. Internal types reachable from an exported signature are re-exported as Go type aliases, which `TestPublicAPIHasNoUnaliasedInternalTypes` enforces, and every exported `With*` option in `internal/engine` must be mirrored here, which `TestEveryEngineOptionIsReExported` enforces (both in `internal/apicheck`).
-- `cmd/cloudstic/` — CLI entry point (`package main`). `command.go` defines the recursive `command` tree (`leaf`/`group`) and its shared dispatcher; `commands.go` holds `commandRegistry()`, the ordered list that is the single source of truth for the command surface: `main.go`'s `runCmd()` dispatches from it, `printUsage()` renders its `COMMANDS` listing from it, and the shell-completion command lists are generated from it. Each command's `run*()` function lives in its own `cmd_*.go` file (e.g. `cmd_backup.go`, `cmd_key.go`). Subcommands: `init`, `backup`, `restore`, `list`, `ls`, `find`, `prune`, `forget`, `diff`, `break-lock`, `key` (with `list`/`add-recovery`/`passwd`), `check`, `cat`, `profile`, `auth`, `store`, `source`, `setup`, `tui`, `completion`. Uses Go's `flag` package (no cobra/viper); `reorderArgs()` in `flags.go` allows flags after positional args. The interactive terminal UI lives in `cmd_tui*.go`.
+- `cmd/cloudstic/` — CLI entry point (`package main`). `command.go` defines the recursive `command` tree (`leaf`/`group`) and its shared dispatcher; `commands.go` holds `commandRegistry()`, the ordered list that is the single source of truth for the command surface: `main.go`'s `runCmd()` dispatches from it, `printUsage()` renders its `COMMANDS` listing from it, and the shell-completion command lists are generated from it. Each command's `run*()` function lives in its own `cmd_*.go` file (e.g. `cmd_backup.go`, `cmd_key.go`). A `cmd_*.go` file holds the command surface — its args, flags, and `run*()`; workflows those commands call into live beside it under the area's own name, so `store_encryption.go` (choosing a store's encryption method and where its secrets live) and `store_health.go` (is a store reachable and initialized, including the AWS SSO re-auth path) sit next to `cmd_store.go` rather than inside it, since `profile new` and `setup` call into both. Subcommands: `init`, `backup`, `restore`, `list`, `ls`, `find`, `prune`, `forget`, `diff`, `break-lock`, `key` (with `list`/`add-recovery`/`passwd`), `check`, `cat`, `profile`, `auth`, `store`, `source`, `setup`, `tui`, `completion`. Uses Go's `flag` package (no cobra/viper); `reorderArgs()` in `flags.go` allows flags after positional args. The interactive terminal UI lives in `cmd_tui*.go`.
   - Commands are free functions taking the dependency container first: `run<Name>(r *runner, ctx, ...)` (and `exec<Name>(r, ctx, ...)` for testable sub-steps). The `runner` struct (`runner.go`) carries `out`/`errOut` writers and the client, so command output is capturable in tests; only I/O primitives (`fail`, `writeJSON`, `openClient`, `prompt*`) remain methods on it.
   - `globalFlags` (`flags.go`) holds only parsed command-line/environment values — it has no methods that build infrastructure. `config.go` defines the resolved configuration types (`clientConfig`, `storeConfig`, `unlockConfig`, …) and the explicit, non-mutating resolution step `resolveClientConfig(g *globalFlags) (clientConfig, error)`, which folds in the selected profile's store without writing back into `g`. Construction takes those config values, never `globalFlags`, so it's unit-testable without going through flag parsing: `storebuild.go` builds the object store, `clientbuild.go` layers a store/keychain/reporter into a repository client, and `keychain.go` builds the keychain/KMS client. `TestGlobalFlagsHasNoConstructionMethods` (`config_test.go`) fails if a construction method is added back onto `globalFlags`.
   - Each command translates its flags into a resolved `pkg/config` value **once**, at the top, and everything below works from that value: `clientConfigFromFlags` for the repository and `backupConfigFromFlags` (`cmd_backup.go`) for what to back up. Both are pure translations — no I/O, no mutation of the args. Nothing downstream reads the flag struct except for the two things a resolved config deliberately does not carry: the output mode, and the store half that `r.openClient` resolves from the selected profile. Source construction and option building are *not* in this package at all — they are `open.Source` and `open.Backup`.
