+++ cmd/cloudstic/cmd_store.go
@@ -4,17 +4,11 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"github.com/cloudstic/cli/pkg/config"
-	"io"
 	"os"
-	"strings"
+
+	"github.com/cloudstic/cli/pkg/config"
 
 	"github.com/cloudstic/cli/pkg/profile"
-
-	cloudstic "github.com/cloudstic/cli"
-	"github.com/cloudstic/cli/internal/ui"
-	"github.com/cloudstic/cli/pkg/secretref"
-	"github.com/cloudstic/cli/pkg/store"
 )
 
 type storeListArgs struct{ profilesFile string }
@@ -338,398 +332,6 @@ func runStoreInit(r *runner, ctx context.Context, a *storeInitArgs) int {
 	return 0
 }
 
-type checkOrInitOptions struct {
-	// configDir is the resolved -config-dir, which the store's secret
-	// references are read relative to. See newSecretResolver.
-	configDir            string
-	allowMissingSecrets  bool
-	warnOnMissingSecrets bool
-	offerInit            bool
-	assumeYes            bool
-}
-
-func checkOrInitStore(r *runner, ctx context.Context, cfg *profile.Config, storeName, profilesFile string, opts checkOrInitOptions) error {
-	s := cfg.Stores[storeName]
-	resolved, err := clientConfigFromProfileStore(s, opts.configDir)
-	if err != nil {
-		if opts.allowMissingSecrets && isSecretNotFoundError(err) {
-			if opts.warnOnMissingSecrets {
-				_, _ = fmt.Fprintf(r.errOut, "Store credentials are configured but not currently available: %v\n", err)
-				_, _ = fmt.Fprintf(r.errOut, "Set required secrets and run: cloudstic store verify %s\n", storeName)
-			}
-			return nil
-		}
-		return fmt.Errorf("could not resolve store credentials: %w", err)
-	}
-	resolved.Unlock.NoPrompt = r.noPrompt
-	raw, err := openStore(ctx, resolved.Store)
-	if err != nil {
-		return fmt.Errorf("could not connect to store: %w", err)
-	}
-
-	// Check if already initialized by looking for the config marker.
-	cfgData, err := raw.Get(ctx, "config")
-	if err == nil && cfgData != nil {
-		_, _ = fmt.Fprintln(r.out, "Store is already initialized and accessible.")
-		// InspectRepo, not LoadRepoConfig: an encrypted repository's marker is
-		// sealed, and this runs before any credentials have been verified —
-		// precisely to decide whether they need to be.
-		repoStatus, cfgErr := cloudstic.InspectRepo(ctx, raw)
-		if cfgErr != nil {
-			return fmt.Errorf("read repository config: %w", cfgErr)
-		}
-		if repoStatus.Encrypted {
-			_, _ = fmt.Fprintln(r.out, "Repository is encrypted; verifying configured credentials...")
-			if err := verifyStoreEncryptionCredentials(ctx, resolved.Unlock, raw); err != nil {
-				return fmt.Errorf("store is initialized, but configured encryption credentials are invalid: %w", err)
-			}
-			_, _ = fmt.Fprintln(r.out, "Encryption credentials are valid.")
-		}
-		return nil
-	}
-
-	_, _ = fmt.Fprintln(r.out, "Store is accessible but not yet initialized.")
-	if !opts.offerInit {
-		return nil
-	}
-	if !opts.assumeYes {
-		yes, promptErr := r.promptConfirm(ctx, "Initialize it now?", true)
-		if promptErr != nil || !yes {
-			return nil
-		}
-	}
-
-	// Check if the store has encryption config.
-	hasEncryption := storeHasExplicitEncryption(s)
-
-	if !hasEncryption {
-		// No encryption configured — init without encryption.
-		result, initErr := cloudstic.InitRepo(ctx, raw, cloudstic.WithInitNoEncryption())
-		if initErr != nil {
-			return fmt.Errorf("init failed: %w", initErr)
-		}
-		printInitResult(r.errOut, result)
-		return nil
-	}
-
-	// Build keychain from the store's encryption settings.
-	// For password-based encryption, the env var must be set for init.
-	// If not set, prompt for the password interactively.
-	kc, err := buildKeychain(ctx, resolved.Unlock)
-	if err != nil {
-		return fmt.Errorf("failed to build keychain: %w", err)
-	}
-
-	var initOpts []cloudstic.InitOption
-	initOpts = append(initOpts, cloudstic.WithInitCredentials(kc))
-	result, err := cloudstic.InitRepo(ctx, raw, initOpts...)
-	if err != nil {
-		return fmt.Errorf("init failed: %w", err)
-	}
-	printInitResult(r.errOut, result)
-	return nil
-}
-
-func checkOrInitStoreWithRecovery(r *runner, ctx context.Context, cfg *profile.Config, storeName, profilesFile string, opts checkOrInitOptions, allowSkip bool) error {
-	for {
-		err := checkOrInitStore(r, ctx, cfg, storeName, profilesFile, opts)
-		if err == nil || !r.canPrompt() {
-			return err
-		}
-
-		s := cfg.Stores[storeName]
-		options := []string{"Retry"}
-		loginOption := awsSSOLoginOption(s, err)
-		if loginOption != "" {
-			options = append(options, loginOption)
-		}
-		if allowSkip {
-			options = append(options, "Skip for now")
-		} else {
-			options = append(options, "Abort")
-		}
-
-		_, _ = fmt.Fprintf(r.errOut, "%v\n", err)
-		picked, promptErr := r.promptSelect(ctx, "Store verification failed", options)
-		if promptErr != nil {
-			return err
-		}
-
-		switch picked {
-		case "Retry":
-			continue
-		case "Skip for now":
-			return nil
-		case "Abort":
-			return err
-		case loginOption:
-			if runErr := runAWSSSOLogin(r, ctx, s); runErr != nil {
-				_, _ = fmt.Fprintf(r.errOut, "AWS SSO login failed: %v\n", runErr)
-			}
-		default:
-			return err
-		}
-	}
-}
-
-// promptEncryptionConfig guides the user through encryption configuration
-// and saves the chosen settings to profiles.yaml. It does not build a keychain
-// or prompt for the actual password — that happens later during init.
-// promptEncryptionConfig walks the user through choosing an encryption method for
-// a store and records the choice in the profiles file.
-//
-// A free function taking the runner: it reads and writes a profiles file and
-// mutates cfg, which is profiles-domain work rather than a runner capability.
-func promptEncryptionConfig(r *runner, ctx context.Context, cfg *profile.Config, storeName, profilesFile, configDir string) {
-	_, _ = fmt.Fprintln(r.out)
-	_, _ = fmt.Fprintln(r.out, "No encryption is configured for this store.")
-
-	options := []string{
-		"Password (recommended for interactive use)",
-		"Platform key (recommended for automation/CI)",
-		"AWS KMS key (enterprise)",
-		"No encryption (not recommended)",
-	}
-	picked, err := r.promptSelect(ctx, "Select encryption method", options)
-	if err != nil {
-		_, _ = fmt.Fprintf(r.errOut, "Failed to select encryption method: %v\n", err)
-		return
-	}
-
-	s, err := configureStoreEncryptionSelection(
-		ctx,
-		cfg.Stores[storeName],
-		storeName,
-		picked,
-		func(ctx context.Context, storeName, secretLabel, defaultEnvName, defaultAccount string) (string, error) {
-			return promptSecretReference(r, ctx, configDir, storeName, secretLabel, defaultEnvName, defaultAccount)
-		},
-		r.promptLine,
-		r.out,
-	)
-	if err != nil {
-		_, _ = fmt.Fprintf(r.errOut, "%v\n", err)
-		return
-	}
-	if picked == options[3] {
-		return
-	}
-
-	// Save updated store config.
-	cfg.Stores[storeName] = s
-	if saveErr := profile.Save(profilesFile, cfg); saveErr != nil {
-		_, _ = fmt.Fprintf(r.errOut, "Warning: could not save encryption settings: %v\n", saveErr)
-	}
-}
-
-func configureStoreEncryptionSelection(
-	ctx context.Context,
-	s profile.Store,
-	storeName, picked string,
-	promptSecretRef func(context.Context, string, string, string, string) (string, error),
-	promptLine func(context.Context, string, string) (string, error),
-	out io.Writer,
-) (profile.Store, error) {
-	switch picked {
-	case "Password (recommended for interactive use)":
-		secretRef, err := promptSecretRef(ctx, storeName, "repository password", "CLOUDSTIC_PASSWORD", "password")
-		if err != nil {
-			return s, fmt.Errorf("failed to configure password secret: %w", err)
-		}
-		s.PasswordSecret = secretRef
-		_, _ = fmt.Fprintf(out, "Encryption: password via %s\n", secretRef)
-	case "Platform key (recommended for automation/CI)":
-		secretRef, err := promptSecretRef(ctx, storeName, "platform key (64-char hex)", "CLOUDSTIC_ENCRYPTION_KEY", "encryption-key")
-		if err != nil {
-			return s, fmt.Errorf("failed to configure platform key secret: %w", err)
-		}
-		s.EncryptionKeySecret = secretRef
-		_, _ = fmt.Fprintf(out, "Encryption: platform key via %s\n", secretRef)
-	case "AWS KMS key (enterprise)":
-		arn, err := promptLine(ctx, "KMS key ARN", "")
-		if err != nil || arn == "" {
-			return s, fmt.Errorf("KMS key ARN is required")
-		}
-		s.KMSKeyARN = arn
-		region, _ := promptLine(ctx, "KMS region", "us-east-1")
-		if region != "" {
-			s.KMSRegion = region
-		}
-		_, _ = fmt.Fprintf(out, "Encryption: AWS KMS (%s)\n", arn)
-	case "No encryption (not recommended)":
-		_, _ = fmt.Fprintln(out, "Encryption: none (not recommended)")
-	default:
-		return s, fmt.Errorf("unsupported encryption selection: %s", picked)
-	}
-	return s, nil
-}
-
-// promptSecretReference asks where a store credential should be stored and
-// returns the secret reference for it.
-func promptSecretReference(r *runner, ctx context.Context, configDir, storeName, secretLabel, defaultEnvName, defaultAccount string) (string, error) {
-	return promptSecretReferenceWithFns(
-		ctx,
-		storeName,
-		secretLabel,
-		defaultEnvName,
-		defaultAccount,
-		func(_ context.Context, l string, o []string) (string, error) { return r.promptSelect(ctx, l, o) },
-		func(ctx context.Context, l, d string) (string, error) { return r.promptLine(ctx, l, d) },
-		func(_ context.Context, s string) (string, error) { return r.promptSecret(ctx, s) },
-		os.LookupEnv,
-		newSecretResolver(configDir),
-	)
-}
-
-func promptSecretReferenceWithFns(
-	ctx context.Context,
-	storeName, secretLabel, defaultEnvName, defaultAccount string,
-	promptSelect func(context.Context, string, []string) (string, error),
-	promptLine func(context.Context, string, string) (string, error),
-	promptSecret func(context.Context, string) (string, error),
-	lookupEnv func(string) (string, bool),
-	resolver *secretref.Resolver,
-) (string, error) {
-	writableBackends := resolver.WritableBackends()
-	nativeRef := func(backend secretref.WritableBackend) (string, error) {
-		ref := backend.DefaultRef(storeName, defaultAccount)
-		exists, err := resolver.Exists(ctx, ref)
-		if err != nil {
-			return "", err
-		}
-		if exists {
-			return ref, nil
-		}
-		secretValue, err := promptSecret(ctx, "Secret value")
-		if err != nil {
-			return "", err
-		}
-		if secretValue == "" {
-			return "", fmt.Errorf("secret value cannot be empty")
-		}
-		if err := resolver.Store(ctx, ref, secretValue); err != nil {
-			return "", err
-		}
-		return ref, nil
-	}
-
-	if len(writableBackends) > 0 {
-		options := []string{"Environment variable (env://)"}
-		backendByOption := map[string]secretref.WritableBackend{}
-		for _, backend := range writableBackends {
-			option := fmt.Sprintf("%s (%s://)", backend.DisplayName(), backend.Scheme())
-			options = append(options, option)
-			backendByOption[option] = backend
-		}
-		picked, err := promptSelect(
-			ctx,
-			fmt.Sprintf("Where should %s be stored?", secretLabel),
-			options,
-		)
-		if err != nil {
-			return "", err
-		}
-		if backend, ok := backendByOption[picked]; ok {
-			return nativeRef(backend)
-		}
-	}
-
-	envName, err := promptLine(ctx, "Env var name", defaultEnvName)
-	if err != nil {
-		return "", err
-	}
-	if _, ok := lookupEnv(envName); !ok && len(writableBackends) > 0 {
-		options := []string{"Keep environment variable reference (env://)"}
-		backendByOption := map[string]secretref.WritableBackend{}
-		for _, backend := range writableBackends {
-			option := fmt.Sprintf("Store in %s instead (%s://)", backend.DisplayName(), backend.Scheme())
-			options = append(options, option)
-			backendByOption[option] = backend
-		}
-		picked, err := promptSelect(
-			ctx,
-			fmt.Sprintf("Environment variable %q is not set in this shell", envName),
-			options,
-		)
-		if err != nil {
-			return "", err
-		}
-		if backend, ok := backendByOption[picked]; ok {
-			return nativeRef(backend)
-		}
-	}
-	return envRef(envName), nil
-}
-
-func isSecretNotFoundError(err error) bool {
-	var refErr *secretref.Error
-	if errors.As(err, &refErr) {
-		return refErr.Kind == secretref.KindNotFound
-	}
-	return false
-}
-
-func verifyStoreEncryptionCredentials(ctx context.Context, cfg unlockConfig, raw store.ObjectStore) error {
-	kc, err := buildKeychain(ctx, cfg)
-	if err != nil {
-		return fmt.Errorf("build keychain: %w", err)
-	}
-	_, err = cloudstic.NewClient(ctx, raw,
-		cloudstic.WithKeychain(kc),
-		cloudstic.WithReporter(ui.NewNoOpReporter()),
-	)
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
-func awsSSOLoginOption(s profile.Store, err error) string {
-	if !isAWSExpiredAuthError(err) {
-		return ""
-	}
-	if s.S3Profile != "" {
-		return fmt.Sprintf("Run aws sso login --profile %s", s.S3Profile)
-	}
-	return "Run aws sso login"
-}
-
-func isAWSExpiredAuthError(err error) bool {
-	if err == nil {
-		return false
-	}
-	msg := strings.ToLower(err.Error())
-	for _, needle := range []string{
-		"sso session has expired",
-		"sso session is expired",
-		"sso session is invalid",
-		"ssoproviderinvalidtoken",
-		"token has expired and refresh failed",
-		"the security token included in the request is expired",
-		"expiredtoken",
-		"expired token",
-		"invalid security token",
-	} {
-		if strings.Contains(msg, needle) {
-			return true
-		}
-	}
-	return false
-}
-
-func runAWSSSOLogin(r *runner, ctx context.Context, s profile.Store) error {
-	args := []string{"sso", "login"}
-	if s.S3Profile != "" {
-		args = append(args, "--profile", s.S3Profile)
-	}
-	_, _ = fmt.Fprintf(r.errOut, "Running: aws %s\n", strings.Join(args, " "))
-	runFn := r.runInteractiveCmd
-	if runFn == nil {
-		runFn = defaultRunInteractiveCmd
-	}
-	return runFn(ctx, r.stdin, r.out, r.errOut, "aws", args...)
-}
-
 func hasStoreNewOverrideFlags(g *globalFlags) bool {
 	for name, origin := range g.origins {
 		if origin != originFlag {
@@ -750,19 +352,6 @@ func existingStoreInteractivePlan(canPrompt, hasOverrides, hasEncryption bool) (
 		return false, false
 	}
 	return true, hasEncryption
-}
-
-func envRef(name string) string {
-	name = strings.TrimSpace(name)
-	if name == "" {
-		return ""
-	}
-	return "env://" + name
-}
-
-func storeHasExplicitEncryption(s profile.Store) bool {
-	return s.PasswordSecret != "" || s.EncryptionKeySecret != "" || s.RecoveryKeySecret != "" ||
-		s.KMSKeyARN != ""
 }
 
 // storeCommand declares the `store` command group.
